Crate whitespacers

Structs

Interpreter
A whitespace interpreter. This struct provides various strategies for interpreting a whitespace program.

Options
These are bitflag options to alter the behaviour of the interpreter.

Program
This struct contains the internal representation of a whitespace program.

WsError
The common error type returned from all whitespacers execution functions.

Enums

WsErrorKind
Simple information on what kind of error occurred.

Constants

IGNORE_OVERFLOW
Use wrapping arithmetric instead of detecting overflow and falling back

NO_FALLBACK
Don't fall back to biginteger interpretation on overflow, instead return an overflow error.

NO_IMPLICIT_EXIT
Raise an error on hitting the end of the program instead of cleanly exiting

UNCHECKED_HEAP
When an unknown key used to get an item from the heap, return 0 instead of causing an error.

Functions

debug_compile
Returns a buffer containing the output of the compilation process of the specified program. This is mainly useful for debugging and optimizing the performance of the JIT compiler.