Enum whitespacers::WsErrorKind

pub enum WsErrorKind {
    ParseError(usize, usize, usize),
    StackError,
    KeyError,
    InvalidIndex,
    CallStackError,
    DivisionError,
    IOError,
    RuntimeParseError,
    Overflow,
    InumOverflow(isize, BigInt),
}

Simple information on what kind of error occurred.

Variants

ParseError(usize, usize, usize)

Compile-time parse error

StackError

The stack was not of the correct size to execute an instruction.

KeyError

A missing key was requested from the heap.

InvalidIndex

The program tried to execute an instruction that doesn't exist (generally caused by control flow hitting the end of the program).

CallStackError

The program tried to return but there was no location to return to on the callstack.

DivisionError

Division or Modulo by zero.

IOError

Something went wrong while trying to read from input or write to output.

RuntimeParseError

The program tried to read a number but no number was given.

Overflow

An overflow occurred during an arithmetric operation. This will normally not be returned unless fallback is disabled.

InumOverflow(isize, BigInt)

An overflow occurred when a number input was requested. This is a bit of a special case, as the state cannot be rewound to before the number was parsed. Therefore, the key where the number will be read to, and the oversized integer that was parsed are returned in the error, and the location at which the error occurred is set to be the operation after the failed inum operation. Again, this will not be returned unless fallback is disabled.