Struct whitespacers::Program

pub struct Program { /* fields omitted */ }

This struct contains the internal representation of a whitespace program.

Methods

impl Program

fn strip(&mut self)

Remove source location information from the program.

impl Program

fn parse(code: Vec<u8>) -> Result<Program, WsError>

Parse a program written in whitespace to a format suitable for execution.

fn dump(&self) -> Vec<u8>

Serialize the internal representation back into a whitespace program.

impl Program

fn disassemble(&self) -> String

Disassemble a program into a human-readable whitespace assembly.

fn assemble(source: String) -> Result<Program, WsError>

Parse a program written in whitespace assembly into a program.